feat(xmtp_mls): shared RearmChannel; DisappearingChannels becomes its alias
